[HDReflectionProbeEditor] Refactored UI for reflection probe
