+++ ScriptableRenderPipeline/HDRenderPipeline/HDRP/Editor/Lighting/SerializedHDReflectionProbe.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4e4bdea98ec8e664b95e5c89aa680636
+guid: bd5f27599094ee7479eb849f9c26c187
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
